Move AvailabilityComponent under Developers module
