+++ app/components/developer_card_component.html.erb
@@ -11,7 +11,7 @@
         </h2>
 
         <div class="hidden sm:inline-flex ml-auto flex-row-reverse shrink-0 items-center text-gray-400">
-          <%= render AvailabilityComponent.new(developer: developer) %>
+          <%= render Developers::AvailabilityComponent.new(developer: developer) %>
         </div>
       </div>
