+++ app/components/developers/availability_component.html.erb
