+++ config/locales/en.yml
@@ -97,10 +97,6 @@ en:
         destroyed: Transaction deleted
       update:
         updated: Transaction updated
-  availability_component:
-    in_future_html: Available in <time datetime="%{date}">%{date_in_words}</time>
-    now_html: Available now
-    unspecified_html: Currently unavailable
   blocks:
     create:
       notice: "%{other_recipient} was blocked."
@@ -193,6 +189,10 @@ en:
     work_preference:
       title: Work preference
   developers:
+    availability_component:
+      in_future_html: Available in <time datetime="%{date}">%{date_in_words}</time>
+      now_html: Available now
+      unspecified_html: Currently unavailable
     create:
       created: Your profile was added!
     details:
